Move semester operations to its own class
