+++ sage-ios/SAGE/SAGE/AdminOperations.swift
@@ -243,30 +243,7 @@ class AdminOperations {
         }
     }
     
-    static func startSemester(semester: Semester, completion: (Semester) -> Void, failure: (String) -> Void) {
-        let params = [
-            SemesterConstants.kStartDate: semester.stringFromStartDate(),
-            SemesterConstants.kTerm: semester.term.rawValue
-        ]
-        
-        BaseOperation.manager().POST(StringConstants.kEndpointStartSemester, parameters: params, success: { (operation, data) -> Void in
-            let semesterDict = (data as! [String: AnyObject])["semester"] as! [String: AnyObject]
-            let createdSemester = Semester(propertyDictionary: semesterDict)
-            completion(createdSemester)
-            }) { (operation, error) -> Void in
-                failure("Could not start semester.")
-        }
-    }
-    
-    static func endSemester(completion: () -> Void, failure: (String) -> Void) {
-        // GET CACHED semester
-        let semester = Semester()
-        BaseOperation.manager().POST(StringConstants.kEndpointEndSemester(semester.id), parameters: nil, success: { (operation, data) -> Void in
-            completion()
-            }) { (operation, error) -> Void in
-                failure("Could not end semester.")
-        }
-    }
+
     
     static func createAnnouncement(announcement: Announcement, completion: (Announcement) -> Void, failure: (String) -> Void) {
         let manager = BaseOperation.manager()
